move display_tasks to utils
