+++ pyproject.toml
@@ -45,7 +45,6 @@ python-telegram-bot = {extras = ["rate-limiter"], version = "^20.2"}
 structlog = "^23.1.0"
 asgi-correlation-id = "^4.2.0"
 rich = "^13.3.5"
-pyngrok = "^6.0.0"
 
 
 [tool.poetry.group.dev.dependencies]
