+++ src/core/utils.py
@@ -1,7 +1,22 @@
+import locale
 import sys
 from functools import wraps
 
 from src.settings import settings
+
+locale.setlocale(locale.LC_ALL, "ru_RU.UTF-8")
+
+
+def display_tasks(task):
+    deadline = task.deadline.strftime("%d.%m.%y")
+    bonus_link = "https://help.procharity.ru/article/10053"
+    return (
+        f"От фонда: {task.name_organization}\n\n"
+        f"Категория: {task.category.name}\n\n"
+        f"Срок: {deadline}\n"
+        f"Бонусы: <a href='{bonus_link}'>{task.bonus * '💎'}</a>\n"
+        f"<a href='<Сделать ссылку на задание>'>{'Посмотреть задание'}</a>"
+    )
 
 
 def auto_commit(func):
